+++ src/DotNetOutdated/ReportingExtensions.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 
 namespace DotNetOutdated
@@ -8,9 +9,15 @@ namespace DotNetOutdated
         public static int[] DetermineColumnWidths(this List<ReportedPackage> packages)
         {
             List<int> columnWidths = new List<int>();
-            columnWidths.Add(packages.Select(p => p.Name).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length);
-            columnWidths.Add(packages.Select(p => p.ReferencedVersion?.ToString() ?? Constants.Reporting.UnknownValue).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length);
-            columnWidths.Add(packages.Select(p => p.LatestVersion?.ToString() ?? Constants.Reporting.UnknownValue).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length);
+            columnWidths.Add(Math.Max(
+                packages.Select(p => p.Name).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length, 
+                Constants.Reporting.Headers.PackageName.Length));
+            columnWidths.Add(Math.Max(
+                packages.Select(p => p.ReferencedVersion?.ToString() ?? Constants.Reporting.UnknownValue).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length,
+                Constants.Reporting.Headers.ReferencedVersion.Length));
+            columnWidths.Add(Math.Max(
+                packages.Select(p => p.LatestVersion?.ToString() ?? Constants.Reporting.UnknownValue).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length,
+                Constants.Reporting.Headers.LatestVersion.Length));
 
             return columnWidths.ToArray();
         }